packages/scrape_squad/realtime_analysis/requirements.txt

phidata==2.6.1
langchain==0.3.10
openai==1.57.0
tweepy==4.14.0
newsapi-python==0.2.7
pymongo==4.10.1
neo4j-driver==5.27.0
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4

feedparser==6.0.11

SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0

plotly==5.19.0
kaleido==0.2.1
nbformat==5.9.2
